Add files via upload
Customer churning is a serious problem for telecom companies as it is more expensive to acquire new customers than it is to keep existing ones from leaving. Telecom wants to retain existing customers by providing suitable intervention to the customers who are likely to churn but needs evidence proving that the targeted approach is accurate. To address the churning problem, me and my team first examined the churning problem across the firm using descriptive analytics. Secondly, we built a logistic regression model to assess the predictive accuracy of the targeted approach